Compile RowCoderImpl with Cython.
This simply moves the implementation so it can be compiled.

Before:
tiny_row, RowCoder, 1000 element(s)    : p. element median time cost: 9.01306e-06 sec, relative std: 1.92%
large_row, RowCoder, 1000 element(s)   : p. element median time cost: 2.85555e-05 sec, relative std: 3.01%
nullable_row, RowCoder, 1000 element(s): p. element median time cost: 4.59391e-05 sec, relative std: 2.02%
diverse_row, RowCoder, 1000 element(s) : p. element median time cost: 2.67515e-05 sec, relative std: 5.12%

After:
tiny_row, RowCoder, 1000 element(s)    : p. element median time cost: 6.6539e-06 sec, relative std: 6.37%
large_row, RowCoder, 1000 element(s)   : p. element median time cost: 2.29471e-05 sec, relative std: 3.08%
nullable_row, RowCoder, 1000 element(s): p. element median time cost: 3.26155e-05 sec, relative std: 3.15%
diverse_row, RowCoder, 1000 element(s) : p. element median time cost: 2.2439e-05 sec, relative std: 2.31%
