+++ sdks/python/apache_beam/coders/coder_impl.py
@@ -24,7 +24,9 @@ allow conditional (compiled/pure) implementations, which can be used to
 encode many elements with minimal overhead.
 
 This module may be optionally compiled with Cython, using the corresponding
-coder_impl.pxd file for type hints.
+coder_impl.pxd file for type hints.  In particular, because CoderImpls are
+never pickled and sent across the wire (unlike Coders themselves) the workers
+can use compiled Impls even if the main program does not (or vice versa).
 
 For internal use only; no backwards-compatibility guarantees.
 """
@@ -35,6 +37,7 @@ import itertools
 import json
 import logging
 import pickle
+from array import array
 from io import BytesIO
 from typing import TYPE_CHECKING
 from typing import Any
@@ -53,9 +56,11 @@ import dill
 from fastavro import parse_schema
 from fastavro import schemaless_reader
 from fastavro import schemaless_writer
+import numpy as np
 
 from apache_beam.coders import observable
 from apache_beam.coders.avro_record import AvroRecord
+from apache_beam.typehints.schemas import named_tuple_from_schema
 from apache_beam.utils import proto_utils
 from apache_beam.utils import windowed_value
 from apache_beam.utils.sharded_key import ShardedKey
@@ -1571,3 +1576,99 @@ class TimestampPrefixingWindowCoderImpl(StreamCoderImpl):
     estimated_size += TimestampCoderImpl().estimate_size(value)
     estimated_size += self._window_coder_impl.estimate_size(value, nested)
     return estimated_size
+
+
+class RowCoderImpl(StreamCoderImpl):
+  """For internal use only; no backwards-compatibility guarantees."""
+  SIZE_CODER = VarIntCoderImpl()
+  NULL_MARKER_CODER = BytesCoderImpl()
+
+  def __init__(self, schema, components):
+    self.schema = schema
+    self.constructor = named_tuple_from_schema(schema)
+    self.encoding_positions = list(range(len(self.schema.fields)))
+    if self.schema.encoding_positions_set:
+      # should never be duplicate encoding positions.
+      enc_posx = list(
+          set(field.encoding_position for field in self.schema.fields))
+      if len(enc_posx) != len(self.schema.fields):
+        raise ValueError(
+            f'''Schema with id {schema.id} has encoding_positions_set=True,
+            but not all fields have encoding_position set''')
+      self.encoding_positions = list(
+          field.encoding_position for field in self.schema.fields)
+    self.encoding_positions_argsort = np.argsort(self.encoding_positions)
+    self.components = list(
+        components[self.encoding_positions.index(i)].get_impl()
+        for i in self.encoding_positions)
+    self.has_nullable_fields = any(
+        field.type.nullable for field in self.schema.fields)
+
+  def encode_to_stream(self, value, out, nested):
+    nvals = len(self.schema.fields)
+    self.SIZE_CODER.encode_to_stream(nvals, out, True)
+    attrs = [getattr(value, f.name) for f in self.schema.fields]
+
+    words = array('B')
+    if self.has_nullable_fields:
+      nulls = [attr is None for attr in attrs]
+      if any(nulls):
+        words = array('B', itertools.repeat(0, (nvals + 7) // 8))
+        for i, is_null in enumerate(nulls):
+          words[i // 8] |= is_null << (i % 8)
+
+    self.NULL_MARKER_CODER.encode_to_stream(words.tobytes(), out, True)
+
+    for i in self.encoding_positions_argsort:
+      if attrs[i] is None:
+        if not self.schema.fields[i].type.nullable:
+          raise ValueError(
+              "Attempted to encode null for non-nullable field \"{}\".".format(
+                  self.schema.fields[i].name))
+        continue
+      self.components[i].encode_to_stream(attrs[i], out, True)
+
+  def decode_from_stream(self, in_stream, nested):
+    nvals = self.SIZE_CODER.decode_from_stream(in_stream, True)
+    words = array('B')
+    words.frombytes(self.NULL_MARKER_CODER.decode_from_stream(in_stream, True))
+
+    if words:
+      nulls = [
+          0 if i // 8 >= len(words) else ((words[i // 8] >> (i % 8)) & 0x01)
+          for i in range(nvals)]
+    else:
+      nulls = itertools.repeat(False, nvals)
+
+    # If this coder's schema has more attributes than the encoded value, then
+    # the schema must have changed. Populate the unencoded fields with nulls.
+    if len(self.components) > nvals:
+      nulls = itertools.chain(
+          nulls, itertools.repeat(True, len(self.components) - nvals))
+
+    # Note that if this coder's schema has *fewer* attributes than the encoded
+    # value, we just need to ignore the additional values, which will occur
+    # here because we only decode as many values as we have coders for.
+
+    sorted_components = [
+        None if is_null else self.components[c].decode_from_stream(
+            in_stream, True) for c,
+        is_null in zip(self.encoding_positions_argsort, nulls)
+    ]
+
+    return self.constructor(
+        *[sorted_components[i] for i in self.encoding_positions])
+
+
+class LogicalTypeCoderImpl(StreamCoderImpl):
+  def __init__(self, logical_type, representation_coder):
+    self.logical_type = logical_type
+    self.representation_coder = representation_coder.get_impl()
+
+  def encode_to_stream(self, value, out, nested):
+    return self.representation_coder.encode_to_stream(
+        self.logical_type.to_representation_type(value), out, nested)
+
+  def decode_from_stream(self, in_stream, nested):
+    return self.logical_type.to_language_type(
+        self.representation_coder.decode_from_stream(in_stream, nested))
